Move FileRegion drain-loop overshoot test to SocketFileRegionTest
Promote testFileRegionDrainStopsAtCompletion and its
OvershootDetectingFileRegion fixture from IoUringSocketFileRegionTest to
the SocketFileRegionTest base class so every transport implementation
(NIO, epoll, kqueue, io_uring and any future addition) is exercised by
the same regression detector.

Drop the io_uring-specific guards along the way: the chunk-size and
"client must be io_uring" assumptions no longer apply once the test
lives in the base class, and all other transports already short-circuit
on transferred() >= count() before invoking transferTo(), so they pass
the assertion as written.
